fix: harden session hook ordering and ownership
